backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.6.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
email-validator>=2.1.0
bcrypt==4.0.1
psycopg2-binary>=2.9.0
openai>=1.0.0
aiofiles>=23.0.0
